Working KTX2 image loading.

+++ LimbicEngine/LimbicTypes.h
@@ -12,7 +12,7 @@ typedef unsigned short uint16;
 typedef short int16;
 typedef unsigned int uint32;
 typedef int int32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 typedef long int64;
 
 struct STexture
